+++ .travis.yml
@@ -1,0 +1,10 @@
+language: python
+python:
+  - "3.7"
+install:
+  - pip install pipenv
+  - pipenv install --system --deploy --ignore-pipfile
+script:
+  - pipenv run pytest
+notifications:
+  email: false
