+++ Pipfile
@@ -12,6 +12,8 @@ waitress = "*"
 whitenoise = "*"
 Django = "*"
 django-otp = "*"
+pytest = "*"
+pytest-django = "*"
 
 [requires]
 python_version = "3.7"
