+++ liquidcore/home/templates/homepage.html
@@ -5,7 +5,7 @@
 
 <p>
   {% if request.user.is_authenticated %}
-    {{ request.user }}
+    <span class="username">{{ request.user }}</span>
     <a href="/accounts/logout/?next=/">[logout]</a>
     {% if request.user.is_staff %}
       <a href="/admin/">[admin]</a>
